backend/requirements.txt
fastapi
uvicorn
python-multipart
groq
pdfplumber
pydanticgroq
python-dotenv
pdfplumber
edge-tts
